Mistery solved
I looked over the RELAYING ACTIONS sections of the book and didn't append an sendAction(for:) statement to imageButtonTapped(_ sender:) so that the moodSelectionChanged(_ sender:) in MoodSelectionViewController will not be called and currentMood will not be updated. This caused the addMoodButton will not be updated in property observer in var currentMood.